requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==2.5.3
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.9.3
entrypoints==0.3
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.8.2
pycountry==18.12.8
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.2
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
wincertstore==0.2
